[MMS] : doods: updated tool improved login steps
