+++ Doods.Openmedivault/Doods.Openmediavault.Mobile.Std/Doods.Openmediavault.Mobile.Std.csproj
@@ -7,7 +7,7 @@
   <ItemGroup>
     <PackageReference Include="Asmodat.Standard.SSH.NET" Version="1.0.0.1" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <PackageReference Include="Xamarin.Forms" Version="4.7.0.968" />
+    <PackageReference Include="Xamarin.Forms" Version="4.8.0.1269" />
   </ItemGroup>
 
   <ItemGroup>
