+++ src/Android/Doods.Xam.MonitorMyServer.Android.csproj
@@ -98,22 +98,22 @@
       <Version>5.2.0</Version>
     </PackageReference>
     <PackageReference Include="AutoMapper">
-      <Version>9.0.0</Version>
+      <Version>10.0.0</Version>
     </PackageReference>
     <PackageReference Include="MarcTron.Admob">
-      <Version>1.5.6</Version>
+      <Version>1.5.7</Version>
     </PackageReference>
     <PackageReference Include="Microsoft.AppCenter">
-      <Version>3.2.2</Version>
+      <Version>3.4.0</Version>
     </PackageReference>
     <PackageReference Include="Microsoft.AppCenter.Analytics">
-      <Version>3.2.2</Version>
+      <Version>3.4.0</Version>
     </PackageReference>
     <PackageReference Include="Microsoft.AppCenter.Crashes">
-      <Version>3.2.2</Version>
+      <Version>3.4.0</Version>
     </PackageReference>
     <PackageReference Include="NLog">
-      <Version>4.7.2</Version>
+      <Version>4.7.3</Version>
     </PackageReference>
     <PackageReference Include="Plugin.Fingerprint">
       <Version>2.1.1</Version>
@@ -122,13 +122,13 @@
       <Version>2.0.0.3</Version>
     </PackageReference>
     <PackageReference Include="SkiaSharp">
-      <Version>1.68.3</Version>
+      <Version>2.80.1</Version>
     </PackageReference>
     <PackageReference Include="SkiaSharp.Views">
-      <Version>1.68.3</Version>
+      <Version>2.80.1</Version>
     </PackageReference>
     <PackageReference Include="SkiaSharp.Views.Forms">
-      <Version>1.68.3</Version>
+      <Version>2.80.1</Version>
     </PackageReference>
     <PackageReference Include="Xam.Plugin.DeviceInfo">
       <Version>4.1.1</Version>
@@ -161,10 +161,10 @@
       <Version>2.4.11.982</Version>
     </PackageReference>
     <PackageReference Include="Xamarin.Forms">
-      <Version>4.7.0.968</Version>
+      <Version>4.8.0.1269</Version>
     </PackageReference>
     <PackageReference Include="Xamarin.Forms.PancakeView">
-      <Version>1.4.2</Version>
+      <Version>2.1.0.714</Version>
     </PackageReference>
     <PackageReference Include="Xamarin.Forms.Svg">
       <Version>1.0.3</Version>
