+++ src/Android/Resources/drawable/splash_screen.xml
@@ -3,8 +3,11 @@
   <item>
     <color android:color="@color/splash_background"/>
   </item>
-  <item>
-    <bitmap
+  <item android:height="300dp" android:gravity="fill" android:color="@color/colorAccent">
+  <bitmap
+    
+      android:layout_width="250dp"
+      android:layout_height="250dp"
       android:src="@mipmap/ic_launcher"
       android:tileMode="disabled"
       android:gravity="center"/>
